user edit page + buy logics

+++ database/migrations/2021_02_11_214528_create_orders_table.php
@@ -20,7 +20,7 @@ class CreateOrdersTable extends Migration
             $table->bigInteger('product_id')->unsigned();
             $table->foreign('product_id')->references('id')->on('products');
             $table->bigInteger('quantity');
-            $table->bigInteger('buy_price');
+            $table->float('buy_price');
             $table->timestamps();
         });
     }
